requirements.txt
asn1crypto==0.24.0
cryptography==2.1.4
enum34==1.1.6
idna==2.6
ipaddress==1.0.17
keyring==10.6.0
keyrings.alt==3.0
Pillow==5.2.0
PyAutoGUI==0.9.38
pycrypto==2.6.1
PyMsgBox==1.0.6
PyScreeze==0.1.18
python-xlib==0.20
PyTweening==1.0.3
pyxdg==0.25
SecretStorage==2.3.1
six==1.11.0
